fix(interruptions): fix interruption duration overflow problems
Interruptions were using ints for durations and that was creating overflow problems such
as displaying incorrect durations or just crashing the app (when overflowing to negative).

+++ app/src/debug/java/com/rbraithwaite/sleeptarget/test_utils/test_data/builders/InterruptionBuilder.java
@@ -33,7 +33,7 @@ public class InterruptionBuilder
 
     private int mId;
     private Date mStart;
-    private int mDuration;
+    private long mDuration;
     private String mReason;
 
 //*********************************************************
